+++ README.md
@@ -2,8 +2,14 @@ TODO
 
 \*\* frontend
 
-- fix custom express server
+- a check
+  request-log
+  l:177
+  l:192
+  l:113
+  l:391
 
-\*\*backend
+  > > > > > > > l:434
+  > > > > > > > \*\*backend
 
 - créer controller form api & gateway
