+++ backend/package.json
@@ -7,6 +7,7 @@
   "dependencies": {
     "express": "^4.16.4",
     "isomorphic-unfetch": "^3.0.0",
+    "request": "^2.88.0",
     "nodemon": "^1.18.10"
   },
   "main": "gateway.js",
