+++ package.json
@@ -4,7 +4,7 @@
   "scripts": {
     "dev-front": "yarn workspace frontend run dev",
     "dev-back": "yarn workspace backend run dev",
-    "start": "run-p dev-front dev-back"
+    "start": "yarn run dev-front"
   },
   "dependencies": {
     "yarn-run-all": "3.1.1"
